+++ packages/terra-application-layout/CHANGELOG.md
@@ -3,7 +3,11 @@ Changelog
 
 Unreleased
 ----------
+
+5.6.0 - (July 10, 2019)
+------------------
 ### Changed
+* Minor dependency version bump
 * Re-enable missing tests.
 
 5.5.0 - (July 2, 2019)